# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Performance Testing Assistant - E-Commerce API",
    "version": "1.0.0",
    "description": "Complete E-Commerce API designed for AI-powered performance testing. Supports automated business journey discovery, load test generation, and performance bottleneck analysis.",
    "contact": {
      "name": "AI Performance Testing Assistant",
      "email": "support@perf-testing.ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {"name": "Authentication", "description": "User authentication endpoints"},
    {"name": "Products", "description": "Product browsing and search"},
    {"name": "Cart", "description": "Shopping cart management"},
    {"name": "Orders", "description": "Order management and history"},
    {"name": "Admin", "description": "Admin dashboard and management"},
    {"name": "Health", "description": "API health check"}
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user account",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "first_name", "last_name"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string", "format": "password", "minLength": 6},
                  "first_name": {"type": "string"},
                  "last_name": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "User registered successfully"},
          "400": {"description": "Invalid input or email already exists"}
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login and receive JWT token",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {"type": "string", "format": "email"},
                  "password": {"type": "string", "format": "password"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Login successful, returns JWT token"},
          "401": {"description": "Invalid credentials"}
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List products with search, filter, and pagination",
        "operationId": "listProducts",
        "security": [],
        "parameters": [
          {"name": "q", "in": "query", "schema": {"type": "string"}, "description": "Search query"},
          {"name": "category_id", "in": "query", "schema": {"type": "integer"}},
          {"name": "min_price", "in": "query", "schema": {"type": "number"}},
          {"name": "max_price", "in": "query", "schema": {"type": "number"}},
          {"name": "sort_by", "in": "query", "schema": {"type": "string", "enum": ["name", "price", "created_at"]}},
          {"name": "sort_order", "in": "query", "schema": {"type": "string", "enum": ["asc", "desc"]}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "List of products with pagination info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {"type": "array", "items": {"$ref": "#/components/schemas/Product"}},
                    "pagination": {"$ref": "#/components/schemas/Pagination"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/categories": {
      "get": {
        "tags": ["Products"],
        "summary": "List all product categories",
        "security": [],
        "responses": {"200": {"description": "List of categories"}}
      }
    },
    "/api/products/{slug}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product details by slug",
        "security": [],
        "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "Product details with related products"},
          "404": {"description": "Product not found"}
        }
      }
    },
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get current user cart",
        "responses": {"200": {"description": "Cart contents with items and total"}}
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Add item to cart",
        "operationId": "addToCart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["product_id"],
                "properties": {
                  "product_id": {"type": "integer"},
                  "quantity": {"type": "integer", "minimum": 1, "default": 1}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Item added to cart"},
          "400": {"description": "Invalid product or insufficient stock"}
        }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Clear entire cart",
        "responses": {"200": {"description": "Cart cleared"}}
      }
    },
    "/api/cart/items/{itemId}": {
      "put": {
        "tags": ["Cart"],
        "summary": "Update cart item quantity",
        "parameters": [{"name": "itemId", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {"quantity": {"type": "integer", "minimum": 0}}
              }
            }
          }
        },
        "responses": {"200": {"description": "Cart item updated"}}
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove item from cart",
        "parameters": [{"name": "itemId", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {"200": {"description": "Item removed from cart"}}
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get user order history",
        "responses": {"200": {"description": "List of user orders"}}
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create new order from cart",
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["shipping_address"],
                "properties": {
                  "shipping_address": {
                    "type": "object",
                    "properties": {
                      "street": {"type": "string"},
                      "city": {"type": "string"},
                      "state": {"type": "string"},
                      "zip": {"type": "string"},
                      "country": {"type": "string"}
                    }
                  },
                  "payment_method": {
                    "type": "string",
                    "enum": ["credit_card", "debit_card", "paypal"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Order created successfully"},
          "400": {"description": "Cart is empty"}
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order details",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Order details with items"},
          "404": {"description": "Order not found"}
        }
      }
    },
    "/api/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get admin dashboard statistics",
        "responses": {"200": {"description": "Dashboard statistics"}}
      }
    },
    "/api/admin/sales": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get sales report data",
        "responses": {"200": {"description": "Sales report data"}}
      }
    },
    "/api/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check endpoint",
        "security": [],
        "responses": {"200": {"description": "API health status"}}
      }
    },
    "/api/metrics": {
      "get": {
        "tags": ["Health"],
        "summary": "Prometheus metrics endpoint",
        "security": [],
        "responses": {"200": {"description": "Prometheus metrics in text format"}}
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/auth/login"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "slug": {"type": "string"},
          "description": {"type": "string"},
          "price": {"type": "number", "format": "float"},
          "compare_price": {"type": "number", "format": "float"},
          "category_name": {"type": "string"},
          "image_url": {"type": "string"},
          "stock_quantity": {"type": "integer"},
          "is_active": {"type": "boolean"}
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {"type": "integer"},
          "limit": {"type": "integer"},
          "total": {"type": "integer"},
          "total_pages": {"type": "integer"},
          "has_next": {"type": "boolean"},
          "has_prev": {"type": "boolean"}
        }
      }
    }
  }
}